shuffle some files around
